+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/hpo/HpoAnnotationModel.java
@@ -17,7 +17,9 @@ import java.util.regex.Pattern;
  *
  * @author <a href="mailto:peter.robinson@jax.org">Peter Robinson</a>
  * Created by peter on 1/20/2018.
+ * @deprecated to be removed in v3.0.0.
  */
+@Deprecated(forRemoval = true)
 public class HpoAnnotationModel {
   /**
    * The base name of the HPO Annotation file.