+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/hpo/OrphanetXML2HpoDiseaseModelParser.java
@@ -53,7 +53,9 @@ import javax.xml.stream.events.XMLEvent;
  * That is, we extract the Orphanumber and the name and then there is a list of annotations.
  *
  * @author Peter Robinson
+ * @deprecated to be removed in v3.0.0.
  */
+@Deprecated(forRemoval = true)
 public class OrphanetXML2HpoDiseaseModelParser {
   private final static Logger logger = LoggerFactory.getLogger(OrphanetXML2HpoDiseaseModelParser.class);
   /**