+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/io/hpo/HpoaDiseaseDataLoaderDefault.java
@@ -19,12 +19,12 @@ import java.util.stream.Collectors;
 class HpoaDiseaseDataLoaderDefault implements HpoaDiseaseDataLoader {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(HpoaDiseaseDataLoaderDefault.class);
-  // To match a line like `#date: 2021-08-02`
-  private static final Pattern VERSION_PATTERN = Pattern.compile("#date: (?<version>[\\w\\d-]+)$");
+  private static final Pattern VERSION_PATTERN = Pattern.compile("#(version|date): (?<version>[\\w-]+)$");
+  private static final Pattern HPO_VERSION_PATTERN = Pattern.compile("#hpo-version: .*/(?<hpoversion>[^/]+)/hp\\.(json|obo\\.owl|obo|owl)$", Pattern.CASE_INSENSITIVE);
 
-  private final Set<String> databasePrefixes;
+  private final Set<DiseaseDatabase> databasePrefixes;
 
-  HpoaDiseaseDataLoaderDefault(Set<String> prefixes) {
+  HpoaDiseaseDataLoaderDefault(Set<DiseaseDatabase> prefixes) {
     this.databasePrefixes = Objects.requireNonNull(prefixes);
   }
 
@@ -33,18 +33,42 @@ class HpoaDiseaseDataLoaderDefault implements HpoaDiseaseDataLoader {
     BufferedReader reader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
 
     Map<TermId, List<HpoAnnotationLine>> annotationLinesByDiseaseId = new HashMap<>();
+    boolean finishedParsingHeader = false;
     String version = null;
+    String hpoVersion = null;
     for (String line = reader.readLine(); line != null; line = reader.readLine()) {
+      // Comment Lines
       if (line.startsWith("#")) {
-
-        if (version == null) {
-          // We don't have the version yet.
+        if(line.startsWith("#version:") || line.startsWith("#date:")){
           Matcher matcher = VERSION_PATTERN.matcher(line);
-          if (matcher.matches())
+          if (matcher.matches()){
             version = matcher.group("version");
+            LOGGER.debug("Detected HPO annotations version {}", version);
+          } else {
+            LOGGER.warn("Could not parse HPO annotations version from the line {}", line);
+          }
+        } else if(line.toLowerCase().startsWith("#hpo-version:")){
+            Matcher matcher = HPO_VERSION_PATTERN.matcher(line);
+            if (matcher.matches()){
+              hpoVersion = matcher.group("hpoversion");
+              LOGGER.debug("HPO version used to generate the annotations {}", hpoVersion);
+            } else {
+              LOGGER.warn("Could not determine HPO version from the line {}", line);
+            }
         }
-
         continue;
+      }
+
+      // The new format has a header that does not start with `#`
+      if (!finishedParsingHeader) {
+        if (line.startsWith("database_id")) {
+          if (line.split("\t").length != 12) {
+            LOGGER.error("HPOA header column does not have expected column length.");
+            throw new IOException("HPOA header column does not have expected column length.");
+          }
+          finishedParsingHeader = true;
+          continue;
+        }
       }
 
       HpoAnnotationLine annotationLine;
@@ -55,21 +79,18 @@ class HpoaDiseaseDataLoaderDefault implements HpoaDiseaseDataLoader {
         continue;
       }
 
-      if (!databasePrefixes.contains(annotationLine.diseaseId().getPrefix()))
+      if (!databasePrefixes.contains(DiseaseDatabase.fromString(annotationLine.diseaseId().getPrefix())))
         continue; // skip unless we want to keep this database
 
       annotationLinesByDiseaseId.computeIfAbsent(annotationLine.diseaseId(), k -> new ArrayList<>())
         .add(annotationLine);
     }
 
-    if (version == null)
-      LOGGER.warn("HPOA disease data version was not found");
-
     List<HpoaDiseaseData> data = annotationLinesByDiseaseId.entrySet().stream()
       .map(processAnnotationEntry())
       .collect(Collectors.toList());
 
-    return new HpoaDiseaseDataContainer(version, data);
+    return new HpoaDiseaseDataContainer(version, hpoVersion, data);
   }
 
   private static Function<Map.Entry<TermId, List<HpoAnnotationLine>>, HpoaDiseaseData> processAnnotationEntry() {